Synchronization with the source base

+++ WebContent/karuta/js/model/Type_Node.js
@@ -170,7 +170,7 @@ UIFactory["Node"].prototype.displayNode = function(type,root,dest,depth,langcode
 					var proxy_parent = proxies_parent[uuid];
 					if (proxy_parent==dest.substring(8) || dest=='contenu') { // dest = {parentid}
 						proxy_target = true;
-						edit = menu = (proxies_edit[uuid].containsArrayElt(g_userroles) || g_userroles[0]=='designer');
+						edit = menu = (proxies_edit[uuid].containsArrayElt(g_userroles) || g_userroles[0]=='designer') && g_edit;
 					}
 			}
 			//============================== ASMCONTEXT =============================
@@ -187,26 +187,34 @@ UIFactory["Node"].prototype.displayNode = function(type,root,dest,depth,langcode
 			if ($("metadata-wad",data)[0]!=undefined && $($("metadata-wad",data)[0]).attr('help')!=undefined && $($("metadata-wad",data)[0]).attr('help')!=""){
 				if (this.depth>0 || this.nodetype == "asmContext") {
 					var help_text = "";
+					var helplang = 0;
+					var display_help = false;
 					var attr_help = $($("metadata-wad",data)[0]).attr('help');
 					var helps = attr_help.split("/"); // lang1/lang2/...
-					if (attr_help.indexOf("@")>-1) { // lang@fr/lang@en/...
-						for (var j=0; j<helps.length; j++){
-							if (helps[j].indexOf("@"+languages[langcode])>-1)
-								help_text = helps[j].substring(0,helps[j].indexOf("@"));
-						}
-					} else { // lang1/lang2/...
-						help_text = helps[langcode];  // lang1/lang2/...
+					for (var j=0; j<helps.length; j++){
+						if (helps[j].indexOf("@"+languages[langcode])>-1)
+							helplang =j;
 					}
-					var help = " <a href='javascript://' class='popinfo'><span style='font-size:12px' class='fas fa-question-circle'></span></a> ";
-					$("#help_"+uuid).html(help);
-					$(".popinfo").popover({ 
-					    placement : 'right',
-					    container : 'body',
-					    title:karutaStr[LANG]['help-label'],
-					    html : true,
-					    trigger:'click hover',
-					    content: help_text
-					});
+					help_text = helps[helplang].substring(0,helps[helplang].indexOf("@"));
+					if (helps[helplang].indexOf(",")>-1) {
+						var roles = helps[helplang].substring(helps[helplang].indexOf(","));
+						if (roles.indexOf(this.userrole)>-1 || (roles.containsArrayElt(g_userroles) && g_userroles[0]!='designer') || USER.admin || g_userroles[0]=='designer')
+						display_help = true;
+					} else {
+						display_help = true;
+					}
+					if (display_help){
+						var help = " <a href='javascript://' class='popinfo'><span style='font-size:12px' class='fas fa-question-circle'></span></a> ";
+						$("#help_"+uuid).html(help);
+						$(".popinfo").popover({ 
+							placement : 'right',
+							container : 'body',
+							title:karutaStr[LANG]['help-label'],
+							html : true,
+							trigger:'click hover',
+							content: help_text
+						});
+					}
 				}
 			}
 			// ===========================================================================
@@ -230,7 +238,7 @@ UIFactory["Node"].prototype.displayNode = function(type,root,dest,depth,langcode
 					}
 				}
 				//------------ Bubble Map -----------------
-				if (this.semantictag=='bubble_level1') {
+				if (this.semantictag.indexOf('bubble_level1')>-1) {
 					alreadyDisplayed = true;
 					if  (this.seeqrcoderoles.containsArrayElt(g_userroles) || USER.admin || g_userroles[0]=='designer') {
 						var map_info = UIFactory.Bubble.getLinkQRcode(uuid);
@@ -349,7 +357,7 @@ UIFactory["Node"].prototype.displayAsmContext = function (dest,type,langcode,edi
 			this.displayMenus("#menus-"+uuid,langcode);
 	}
 	//----------------delete control on proxy parent ------------
-	if (proxies_delete[uuid]!=undefined && proxies_delete[uuid].containsArrayElt(g_userroles)) {
+	if (edit && proxies_delete[uuid]!=undefined && proxies_delete[uuid].containsArrayElt(g_userroles)) {
 		var html = deleteButton(proxies_nodeid[uuid],"asmContext",undefined,undefined,"UIFactory.Node.reloadUnit",g_portfolioid,null);
 		$("#buttons-"+uuid).html(html);
 	}
@@ -475,7 +483,7 @@ UIFactory["Node"].prototype.displayAsmNode = function(dest,type,langcode,edit,re
 	//-------------- label --------------------------
 	var gotView = false;
 	var label_html = ""
-	if (this.semantictag=='bubble_level1'){
+	if (this.semantictag.indexOf("bubble_level1")>-1){
 		label_html += " "+UICom.structure["ui"][uuid].getBubbleView('std_node_'+uuid);
 		gotView = true;
 	}
@@ -504,7 +512,7 @@ UIFactory["Node"].prototype.displayAsmNode = function(dest,type,langcode,edit,re
 			this.displayMenus("#menus-"+uuid,langcode);
 	}
 	//----------------delete control on proxy parent ------------
-	if (proxies_delete[uuid]!=undefined && proxies_delete[uuid].containsArrayElt(g_userroles)) {
+	if (edit && proxies_delete[uuid]!=undefined && proxies_delete[uuid].containsArrayElt(g_userroles)) {
 		var html = deleteButton(proxies_nodeid[uuid],"asmContext",undefined,undefined,"UIFactory.Node.reloadUnit",g_portfolioid,null);
 		$("#buttons-"+uuid).html(html);
 	}
@@ -990,9 +998,12 @@ UIFactory["Node"].remove = function(uuid,callback,param1,param2)
 {
 	//-------- if function js -------------
 	if (UICom.structure.ui[uuid].js!=undefined && UICom.structure.ui[uuid].js!="") {
-		var elts = UICom.structure.ui[uuid].js.split("/");
-		if (elts[0]=="delete")
-			eval(elts[1]+"(UICom.structure.ui[uuid].node,g_portfolioid)");
+		var fcts = UICom.structure.ui[uuid].js.split(";");
+		for (var i=0;i<fcts.length;i++) {
+			var elts = fcts[i].split("/");
+			if (elts[0]=="delete")
+				eval(elts[1]+"(UICom.structure.ui[uuid].node,g_portfolioid)");
+		}
 	}
 	//---------------------
 	$("#"+uuid,g_portfolio_current).remove();
@@ -1632,7 +1643,7 @@ UIFactory["Node"].selectNode = function(nodeid,node,semtag)
 	var uuid = $(node.node).attr("id");
 	var label = UICom.structure["ui"][uuid].label_node[langcode].text();
 	html += "<option uuid = ''>&nbsp;</option>";
-	if (semtag==null)
+	if (semtag=='')
 		html += "<option uuid = '"+uuid+"'>"+label+"</option>";
 	html += UIFactory["Node"].getSubNodes(node, nodeid, UICom.structure.ui[nodeid].asmtype,semtag);
 	html += "</select>";
@@ -1659,7 +1670,7 @@ UIFactory["Node"].getSubNodes = function(root, idmoved, typemoved,semtag)
 			var label = UICom.structure["ui"][uuid].label_node[langcode].text();
 			var name = UICom.structure["ui"][uuid].asmtype;
 			if (name!='asmContext' && (typemoved != "asmUnit" || name != "asmUnit") && (uuid !=idmoved)){
-				if (semantictag.indexOf("welcome-unit")<0 && (semtag==null || (semtag!=null && semantictag==semtag)))
+				if (semantictag.indexOf("welcome-unit")<0 && (semtag=='' || (semtag!=Z && semantictag==semtag)))
 					html += "<option uuid = '"+uuid+"'>"+label+"</option>";
 				html += UIFactory["Node"].getSubNodes(UICom.structure["tree"][uuid], idmoved, typemoved,semtag);
 			}
@@ -1726,10 +1737,10 @@ UIFactory["Node"].prototype.getButtons = function(dest,type,langcode,inline,dept
 		}
 		if (((this.writenode && (this.moveinroles.containsArrayElt(g_userroles)  || this.moveinroles.indexOf($(USER.username_node).text())>-1)) || USER.admin || g_userroles[0]=='designer') && this.asmtype != 'asmRoot') {
 			var movein = ($(this.metadatawad).attr('movein')==undefined)?'':$(this.metadatawad).attr('movein');
-			if (movein=='')
-				html+= "<span class='button fas fa-random' style='"+menus_color+"' onclick=\"javascript:UIFactory.Node.selectNode('"+this.id+"',UICom.root)\" data-title='"+karutaStr[LANG]["move"]+"' data-toggle='tooltip' data-placement='bottom'></span>";
-			else
-				html+= "<span class='button fas fa-random' style='"+menus_color+"' onclick=\"javascript:UIFactory.Node.selectNode('"+this.id+"',UICom.structure.tree[$('#page').attr('uuid')],'"+movein+"')\" data-title='"+karutaStr[LANG]["move"]+"' data-toggle='tooltip' data-placement='bottom'></span>";
+//			if (movein=='')
+				html+= "<span class='button fas fa-random' style='"+menus_color+"' onclick=\"javascript:UIFactory.Node.selectNode('"+this.id+"',UICom.root,'"+movein+"')\" data-title='"+karutaStr[LANG]["move"]+"' data-toggle='tooltip' data-placement='bottom'></span>";
+//			else
+//				html+= "<span class='button fas fa-random' style='"+menus_color+"' onclick=\"javascript:UIFactory.Node.selectNode('"+this.id+"',UICom.structure.tree[$('#page').attr('uuid')],'"+movein+"')\" data-title='"+karutaStr[LANG]["move"]+"' data-toggle='tooltip' data-placement='bottom'></span>";
 		}
 		//------------- duplicate node buttons ---------------
 		if ( (g_userroles[0]=='designer' && this.asmtype != 'asmRoot') // always duplicate for designer
@@ -1809,13 +1820,15 @@ UIFactory['Node'].reloadUnit = function(uuid,redisplay)
 {
 	if (redisplay == null)
 		redisplay = true;
-	if (redisplay)
+	if (uuid=="" || uuid==null || redisplay)
 		uuid = $("#page").attr('uuid');
 	var parentid = "";
 	if (uuid==g_portfolio_rootid)
 		parentid = g_portfolio_rootid;
 	else
 		parentid = $($(UICom.structure["ui"][uuid].node).parent()).attr('id');
+	if (uuid.indexOf("_")>-1)
+		uuid = uuid.substring(0,uuid.indexOf("_"));
 	$.ajax({
 		async: false,
 		type : "GET",
